Refactor util package
